+++ plugin-maven/CHANGES.md
@@ -3,7 +3,10 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (starting after version `1.27.0`).
 
 ## [Unreleased]
+### Added
+* Add `includeDraft` option, to include draft mutators from composite mutators ([#XXX](https://github.com/diffplug/spotless/pull/XXX))
 ### Changes
+* Bump default `cleanthat` version to latest `2.2` -> `2.3` ([#XXX](https://github.com/diffplug/spotless/pull/XXX))
 * Bump default `cleanthat` version to latest `2.1` -> `2.2` ([#1569](https://github.com/diffplug/spotless/pull/1569))
 
 ## [2.33.0] - 2023-02-10